data recovery, added args on cmds
todo edit - editing must accept words as input on priority
todo rm args
todo edit args
(limit max string lenth %padding.maxs)
todo rm/edit/get/random prompts user input even when no items
todo view
todo version
bin serialization version (for updating the app) // byte alignment
displaying strings can destroy UI

+++ src/types.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef uint8_t pScore;
+typedef long long pScore;
 
 typedef struct TNode{
     unsigned int value;
@@ -62,7 +62,7 @@ typedef struct DynamicTodoList
     todotreeCmpFun sortingFunc;
     long long timeToday;
     uint8_t isAccending; 
-    uint8_t energy; 
+    unsigned int energy; 
 } TodoList;
 
 /*
